+++ src/ska_mid_dish_steering_control/sculib.py
@@ -142,8 +142,9 @@ SPM_SCHEMA_PATH: Final = Path(
 
 TOP_NODE: Final = "Server"
 TOP_NODE_ID: Final = "i=2253"
-SKIP_NODE_ID: Final = "ns=22;s=Trace"
 SERVER_STATUS_ID: Final = "i=2256"
+PHYSICAL_NAMESPACE: Final = ua.Int16(7)
+TRACE_NAMESPACE: Final = ua.Int16(22)
 PLC_PRG: Final = "Logic.Application.PLC_PRG"
 APP_STATE: Final = "Logic.Application.ApplicationState"
 CURRENT_MODE: Final = "System.CurrentMode"
@@ -469,8 +470,10 @@ class SteeringControlUnit:
 
         :raises Exception: If connection to OPC UA server fails.
         """
+        logger.debug("Connecting to server...")
         try:
             self._client = self.connect()
+            logger.debug("Connection successful.")
         except Exception as e:
             msg = (
                 "Cannot connect to the OPC UA server. Please "
@@ -859,6 +862,7 @@ class SteeringControlUnit:
         corresponding ua attributes. Add all found enum types and their values to the
         'opcua_enum_types' dictionary property.
         """
+        logger.debug("Validating Enum types...")
         missing_types = []
         expected_types = [
             "AxisSelectType",
@@ -903,6 +907,8 @@ class SteeringControlUnit:
                 "as expected: %s",
                 str(missing_types),
             )
+        else:
+            logger.debug("Enum types valid.")
 
     # ----------------------
     # Command user authority
@@ -1235,6 +1241,7 @@ class SteeringControlUnit:
         This method may raise exceptions related to asyncio operations.
         """
         # Create node dicts of the top node's tree
+        logger.debug("Scanning from %s: %s", TOP_NODE, TOP_NODE_ID)
         (
             self._nodes,
             self._attributes,
@@ -1243,17 +1250,29 @@ class SteeringControlUnit:
         ) = self._check_cache_and_generate_node_dicts(
             top_level_node=self._client.get_node(TOP_NODE_ID),
             top_level_node_name=TOP_NODE,
-            skip_node_id=SKIP_NODE_ID,
+            skip_namespaces=[
+                PHYSICAL_NAMESPACE,
+                TRACE_NAMESPACE,
+                self._parameter_ns_idx,
+            ],
         )
         # Check if the PLC_PRG node exists, and if not, scan it separately and add its
         # nodes, attributes, and commands to the dictionaries. (needed for simulators)
         if PLC_PRG not in self._nodes:
-            plc_prg_node = asyncio.run_coroutine_threadsafe(
-                self._client.nodes.objects.get_child(
-                    [f"{self._ns_idx}:{node}" for node in PLC_PRG.split(".")]
-                ),
-                self.event_loop,
-            ).result()
+            try:
+                plc_prg_node = asyncio.run_coroutine_threadsafe(
+                    self._client.nodes.objects.get_child(
+                        [f"{self._ns_idx}:{node}" for node in PLC_PRG.split(".")]
+                    ),
+                    self.event_loop,
+                ).result()
+            except ua.UaStatusCodeError as e:
+                msg = f"Failed to get child node under Objects at '{PLC_PRG}'."
+                asyncio.run_coroutine_threadsafe(
+                    handle_exception(e, msg), self.event_loop
+                )
+                raise e
+            logger.debug("Scanning from %s: %s", PLC_PRG, plc_prg_node)
             (
                 nodes,
                 attributes,
@@ -1282,6 +1301,7 @@ class SteeringControlUnit:
                 ),
                 self.event_loop,
             ).result()
+            logger.debug("Scanning from %s: %s", top_node_name, parameter_node)
             (
                 self._parameter_nodes,
                 self._parameter_attributes,
@@ -1320,7 +1340,7 @@ class SteeringControlUnit:
         self,
         top_level_node: Node,
         top_level_node_name: str,
-        skip_node_id: str | None = None,
+        skip_namespaces: list[ua.Int16] | None = None,
     ) -> tuple[NodeDict, AttrDict, CmdDict, str]:
         """
         Check for an existing nodes cache to use and generate the dicts accordingly.
@@ -1331,8 +1351,8 @@ class SteeringControlUnit:
 
         :param top_level_node: The top-level node for which to generate dictionaries.
         :param top_level_node_name: Name of the top-level node.
-        :param skip_node_id: Optional ID of a sub-node to skip when generating the
-            dictionaries, defaults to None.
+        :param skip_namespaces: Optional list of namespace indexes to skip when
+            generating the dictionaries, defaults to None.
         :return: A tuple containing dictionaries for nodes, attributes, and commands, as
             well as a string timestamp of when the dicts were generated.
         """
@@ -1355,7 +1375,7 @@ class SteeringControlUnit:
             timestamp = cached_nodes["timestamp"]
         else:
             node_dicts = self.generate_node_dicts_from_server(
-                top_level_node, top_level_node_name, skip_node_id
+                top_level_node, top_level_node_name, skip_namespaces
             )
             timestamp = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
             self._cache_node_ids(cache_file_path, node_dicts[0])
@@ -1383,6 +1403,7 @@ class SteeringControlUnit:
         attributes: AttrDict = {}
         commands: CmdDict = {}
         for node_name, node_details in cache_dict.items():
+            logger.debug("Node from cache: %s", node_name)
             node_id = node_details["opcua_node_str"]
             node_class = node_details["node_class"]
             node: Node = self._client.get_node(node_id)
@@ -1415,7 +1436,7 @@ class SteeringControlUnit:
         self,
         top_level_node: Node,
         top_level_node_name: str,
-        skip_node_id: str | None = None,
+        skip_namespaces: list[ua.Int16] | None = None,
     ) -> tuple[NodeDict, AttrDict, CmdDict]:
         """
         Generate dicts for nodes, attributes, and commands for a given top-level node.
@@ -1428,8 +1449,8 @@ class SteeringControlUnit:
 
         :param top_level_node: The top-level node for which to generate dictionaries.
         :param top_level_node_name: Name of the top-level node.
-        :param skip_node_id: Optional ID of a sub-node to skip when generating the
-            dictionaries, defaults to None.
+        :param skip_namespaces: Optional list of namespace indexes to skip when
+            generating the dictionaries, defaults to None.
         :return: A tuple containing dictionaries for nodes, attributes, and commands.
         """
         logger.info(
@@ -1437,7 +1458,7 @@ class SteeringControlUnit:
             top_level_node_name,
         )
         nodes, attributes, commands = asyncio.run_coroutine_threadsafe(
-            self._get_sub_nodes(top_level_node, top_level_node_name, skip_node_id),
+            self._get_sub_nodes(top_level_node, top_level_node_name, skip_namespaces),
             self.event_loop,
         ).result()
         nodes.update({top_level_node_name: (top_level_node, ua.NodeClass.Object)})
@@ -1489,11 +1510,12 @@ class SteeringControlUnit:
             node_name = node_name.removeprefix(f"{PLC_PRG}.")
         return node_name, parents
 
+    # pylint: disable=too-many-branches
     async def _get_sub_nodes(
         self,
         node: Node,
         top_level_node_name: str,
-        skip_node_id: str | None = None,
+        skip_namespaces: list[ua.Int16] | None = None,
         parent_names: list[str] | None = None,
     ) -> tuple[NodeDict, AttrDict, CmdDict]:
         """
@@ -1501,28 +1523,37 @@ class SteeringControlUnit:
 
         :param node: The node to retrieve sub-nodes from.
         :param top_level_node_name: Name of the top-level node.
-        :param skip_node_id: Optional ID of a sub-node to skip when generating the
-            dictionaries, defaults to None.
+        :param skip_namespaces: Optional list of namespace indexes to skip when
+            generating the dictionaries, defaults to None.
         :param parent_names: List of parent node names (default is None).
         :return: A tuple containing dictionaries of nodes, attributes, and commands.
         """
-        if skip_node_id and skip_node_id == node.nodeid.to_string():
+        logger.debug("Scanning reached %s", node)
+        if skip_namespaces and node.nodeid.NamespaceIndex in skip_namespaces:
+            return {}, {}, {}
+        try:
+            node_class = await node.read_node_class()
+        except (ua.UaStatusCodeError, asyncio.TimeoutError) as e:
+            msg = f"Failed trying to read class of node '{node.nodeid.Identifier}'."
+            asyncio.run_coroutine_threadsafe(handle_exception(e, msg), self.event_loop)
+            return {}, {}, {}
+        try:
+            node_name, parents = await self._generate_full_node_name(
+                node, top_level_node_name, parent_names, node_class
+            )
+        except (ua.UaStatusCodeError, asyncio.TimeoutError) as e:
+            msg = f"Failed trying to read name of node '{node.nodeid.Identifier}'."
+            asyncio.run_coroutine_threadsafe(handle_exception(e, msg), self.event_loop)
+            return {}, {}, {}
+        # Do not add the InputArgument and OutputArgument nodes.
+        if node_name.endswith(".InputArguments") or node_name.endswith(
+            ".OutputArguments"
+        ):
             return {}, {}, {}
 
         nodes: NodeDict = {}
         attributes: AttrDict = {}
         commands: CmdDict = {}
-
-        node_class = await node.read_node_class()
-        node_name, parents = await self._generate_full_node_name(
-            node, top_level_node_name, parent_names, node_class
-        )
-        # Do not add the InputArgument and OutputArgument nodes.
-        if node_name.endswith(".InputArguments") or node_name.endswith(
-            ".OutputArguments"
-        ):
-            return nodes, attributes, commands
-
         nodes[node_name] = (node, node_class)
         # Normal node with children
         # Note: 'ServerStatus' is a standard OPC-UA variable node that has children
@@ -1530,16 +1561,39 @@ class SteeringControlUnit:
             node_class == ua.NodeClass.Object
             or node.nodeid.to_string() == SERVER_STATUS_ID
         ):
-            children = await node.get_children()
-            tasks = [
-                self._get_sub_nodes(child, top_level_node_name, skip_node_id, parents)
-                for child in children
-            ]
-            results = await asyncio.gather(*tasks)
-            for child_nodes, child_attributes, child_commands in results:
-                nodes.update(child_nodes)
-                attributes.update(child_attributes)
-                commands.update(child_commands)
+            try:
+                children = await node.get_children()
+            except ua.UaStatusCodeError as e:
+                msg = f"Server error trying to get children of node '{node_name}'."
+                asyncio.run_coroutine_threadsafe(
+                    handle_exception(e, msg), self.event_loop
+                )
+            except asyncio.TimeoutError as e:
+                msg = (
+                    f"Could not get children of node '{node_name}' in time from server."
+                )
+                asyncio.run_coroutine_threadsafe(
+                    handle_exception(e, msg), self.event_loop
+                )
+            else:
+                tasks = [
+                    self._get_sub_nodes(
+                        child, top_level_node_name, skip_namespaces, parents
+                    )
+                    for child in children
+                ]
+                try:
+                    results = await asyncio.gather(*tasks)
+                except asyncio.TimeoutError as e:
+                    msg = "Gathered async task(s) timed out waiting for result(s)."
+                    asyncio.run_coroutine_threadsafe(
+                        handle_exception(e, msg), self.event_loop
+                    )
+                else:
+                    for child_nodes, child_attributes, child_commands in results:
+                        nodes.update(child_nodes)
+                        attributes.update(child_attributes)
+                        commands.update(child_commands)
         elif node_class == ua.NodeClass.Variable:  # Attribute
             # Check if RO or RW and call the respective creator functions.
             # access_level_set = await node.get_access_level()